# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "List of books documentation API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/books/sortbytitlebottom/": {
            "post": {
                "tags": [
                    "SortController"
                ],
                "summary": "Sort book by title Z-A",
                "operationId": "8613f6d0a848c7b444290ac8ddf93777",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "title": {
                                                "type": "string",
                                                "example": "Some title"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": "Some title"
                                                },
                                                "author": {
                                                    "type": "string",
                                                    "example": "Some author"
                                                },
                                                "published_year": {
                                                    "type": "integer",
                                                    "example": 2025
                                                },
                                                "isbn": {
                                                    "type": "string",
                                                    "example": "Some isbn"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/books/sortbytitletop/": {
            "post": {
                "tags": [
                    "SortController"
                ],
                "summary": "Sort book by title A-Z",
                "operationId": "d149f568cb135add0039688913b5368f",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "title": {
                                                "type": "string",
                                                "example": "Some title"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": "Some title"
                                                },
                                                "author": {
                                                    "type": "string",
                                                    "example": "Some author"
                                                },
                                                "published_year": {
                                                    "type": "integer",
                                                    "example": 2025
                                                },
                                                "isbn": {
                                                    "type": "string",
                                                    "example": "Some isbn"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/books/sortbyyearbottom/": {
            "post": {
                "tags": [
                    "SortController"
                ],
                "summary": "Sort book by year 10-1",
                "operationId": "ec7c7c9820acd31fd3ee35bea7e287ed",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "published_year": {
                                                "type": "integer",
                                                "example": 2025
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": "Some title"
                                                },
                                                "author": {
                                                    "type": "string",
                                                    "example": "Some author"
                                                },
                                                "published_year": {
                                                    "type": "integer",
                                                    "example": 2025
                                                },
                                                "isbn": {
                                                    "type": "string",
                                                    "example": "Some isbn"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/books/sortbyyeartop/": {
            "post": {
                "tags": [
                    "SortController"
                ],
                "summary": "Sort book by year 1-10",
                "operationId": "e11287ef0ff980435c5e5ca7e9e70691",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "published_year": {
                                                "type": "integer",
                                                "example": 2025
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": "Some title"
                                                },
                                                "author": {
                                                    "type": "string",
                                                    "example": "Some author"
                                                },
                                                "published_year": {
                                                    "type": "integer",
                                                    "example": 2025
                                                },
                                                "isbn": {
                                                    "type": "string",
                                                    "example": "Some isbn"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
...
}